requirements.txt
annotated-types==0.7.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
ynab==1.4.0
